Added inputs for MHP
I added a first round of text inputs onto the planscreen. My plan is to then have this data be carried over to a second screen so that we can have it as a test case and proof of concept. We can then use it eventually to further flesh out the MHP entry.